Split bulk-actions tests into one file per action
Replaces the monolithic test/bulk-actions.test.ts with four focused files
under test/bulk-actions/: landing, duplicate, deactivate, reactivate.
